mocks for meta data dao moved into separate file (mock_daos.go -> mock_dao_meta_data.go)
