+++ internal/testutils/mocks/mock_dao_meta_data.go
@@ -1,0 +1,72 @@
+package mocks
+
+import (
+	"fmt"
+
+	m "github.com/RedHatInsights/sources-api-go/model"
+	"github.com/RedHatInsights/sources-api-go/util"
+)
+
+type MockMetaDataDao struct {
+	MetaDatas []m.MetaData
+}
+
+func (a *MockMetaDataDao) List(limit int, offset int, filters []util.Filter) ([]m.MetaData, int64, error) {
+	count := int64(len(a.MetaDatas))
+	return a.MetaDatas, count, nil
+}
+
+func (a *MockMetaDataDao) SubCollectionList(primaryCollection interface{}, limit, offset int, filters []util.Filter) ([]m.MetaData, int64, error) {
+	var appMetaDataList []m.MetaData
+
+	for index, i := range a.MetaDatas {
+		switch object := primaryCollection.(type) {
+		case m.ApplicationType:
+			if i.ApplicationTypeID == object.Id {
+				appMetaDataList = append(appMetaDataList, a.MetaDatas[index])
+			}
+		default:
+			return nil, 0, fmt.Errorf("unexpected primary collection type")
+		}
+	}
+	count := int64(len(appMetaDataList))
+	if count == 0 {
+		return nil, count, util.NewErrNotFound("metadata")
+	}
+
+	return appMetaDataList, count, nil
+}
+
+func (a *MockMetaDataDao) GetById(id *int64) (*m.MetaData, error) {
+	for _, i := range a.MetaDatas {
+		if i.ID == *id {
+			return &i, nil
+		}
+	}
+
+	return nil, util.NewErrNotFound("metadata")
+}
+
+func (a *MockMetaDataDao) Create(src *m.MetaData) error {
+	panic("not implemented") // TODO: Implement
+}
+
+func (a *MockMetaDataDao) Update(src *m.MetaData) error {
+	panic("not implemented") // TODO: Implement
+}
+
+func (a *MockMetaDataDao) Delete(id *int64) error {
+	panic("not implemented") // TODO: Implement
+}
+
+func (md *MockMetaDataDao) GetSuperKeySteps(_ int64) ([]m.MetaData, error) {
+	panic("not implemented")
+}
+
+func (md *MockMetaDataDao) GetSuperKeyAccountNumber(applicationTypeId int64) (string, error) {
+	panic("not implemented!")
+}
+
+func (md *MockMetaDataDao) ApplicationOptedIntoRetry(applicationTypeId int64) (bool, error) {
+	panic("not implemented!")
+}
